refactor: 쿠폰 관리 기능 추가 및 관련 컴포넌트 통합
- CouponsManagement 컴포넌트를 추가하여 쿠폰 관리 UI를 분리
- CouponsForm 컴포넌트를 추가하여 쿠폰 생성 폼 기능 구현
- App.tsx에서 쿠폰 관련 로직을 CouponsManagement로 통합하여 코드 가독성 향상
- types.ts에 CouponForm 타입 추가
