+++ src/basic/components/admin/CouponsManagement.tsx
@@ -1,0 +1,97 @@
+import { Coupon, CouponForm } from '../../../types';
+import Button from '../ui/Button';
+import IconButton from '../ui/IconButton';
+import CouponsForm from './forms/CouponsForm';
+
+interface CouponsManagementProps {
+  coupons: Coupon[];
+  showCouponForm: boolean;
+  couponForm: CouponForm;
+  onCouponDelete: (code: string) => void;
+  onToggleForm: (show: boolean) => void;
+  onCouponSubmit: (e: React.FormEvent) => void;
+  onCouponFormChange: (form: CouponForm) => void;
+  onNotify: (message: string, type: 'error' | 'success' | 'warning') => void;
+}
+
+export default function CouponsManagement({
+  coupons,
+  showCouponForm,
+  couponForm,
+  onCouponDelete,
+  onToggleForm,
+  onCouponSubmit,
+  onCouponFormChange,
+  onNotify,
+}: CouponsManagementProps) {
+  return (
+    <section className='bg-white rounded-lg border border-gray-200'>
+      <div className='p-6 border-b border-gray-200'>
+        <h2 className='text-lg font-semibold'>쿠폰 관리</h2>
+      </div>
+      <div className='p-6'>
+        <div className='grid grid-cols-1 gap-4 sm:grid-cols-2 lg:grid-cols-3'>
+          {coupons.map((coupon) => (
+            <div
+              key={coupon.code}
+              className='relative bg-gradient-to-r from-indigo-50 to-purple-50 rounded-lg p-4 border border-indigo-200'
+            >
+              <div className='flex justify-between items-start'>
+                <div className='flex-1'>
+                  <h3 className='font-semibold text-gray-900'>{coupon.name}</h3>
+                  <p className='text-sm text-gray-600 mt-1 font-mono'>{coupon.code}</p>
+                  <div className='mt-2'>
+                    <span className='inline-flex items-center px-3 py-1 rounded-full text-sm font-medium bg-white text-indigo-700'>
+                      {coupon.discountType === 'amount'
+                        ? `${coupon.discountValue.toLocaleString()}원 할인`
+                        : `${coupon.discountValue}% 할인`}
+                    </span>
+                  </div>
+                </div>
+
+                <IconButton variant='danger' onClick={() => onCouponDelete(coupon.code)}>
+                  <svg className='w-5 h-5' fill='none' stroke='currentColor' viewBox='0 0 24 24'>
+                    <path
+                      strokeLinecap='round'
+                      strokeLinejoin='round'
+                      strokeWidth={2}
+                      d='M19 7l-.867 12.142A2 2 0 0116.138 21H7.862a2 2 0 01-1.995-1.858L5 7m5 4v6m4-6v6m1-10V4a1 1 0 00-1-1h-4a1 1 0 00-1 1v3M4 7h16'
+                    />
+                  </svg>
+                </IconButton>
+              </div>
+            </div>
+          ))}
+
+          <div className='border-2 border-dashed border-gray-300 rounded-lg p-4 flex items-center justify-center hover:border-gray-400 transition-colors'>
+            <Button
+              onClick={() => onToggleForm(!showCouponForm)}
+              className='!text-gray-400 hover:text-gray-600 flex flex-col items-center'
+              variant='ghost'
+            >
+              <svg className='w-8 h-8' fill='none' stroke='currentColor' viewBox='0 0 24 24'>
+                <path
+                  strokeLinecap='round'
+                  strokeLinejoin='round'
+                  strokeWidth={2}
+                  d='M12 4v16m8-8H4'
+                />
+              </svg>
+              <p className='mt-2 text-sm font-medium'>새 쿠폰 추가</p>
+            </Button>
+          </div>
+        </div>
+
+        {showCouponForm && (
+          <CouponsForm
+            couponForm={couponForm}
+            onCouponSubmit={onCouponSubmit}
+            onCouponFormChange={onCouponFormChange}
+            onNotify={onNotify}
+            onToggleForm={onToggleForm}
+          />
+        )}
+      </div>
+    </section>
+  );
+}
